requirements.txt
lenses~=1.1.0
pytest==7.2.0
python-gitlab==3.11.0
PyYAML~=6.0
requests~=2.28.1
setuptools==65.5.0
